--- android/app/src/main/java/com/woowacourse/ody/data/remote/entity/meeting/MeetingsResponse.kt
@@ -1,10 +1,0 @@
-package com.woowacourse.ody.data.remote.entity.meeting
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class MeetingsResponse(
-    @Json(name = "meetings")
-    val meetings: List<MeetingResponse>,
-)
