--- android/app/src/main/java/com/woowacourse/ody/data/remote/entity/notification/NotificationLogResponse.kt
@@ -1,14 +1,0 @@
-package com.woowacourse.ody.data.remote.entity.notification
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class NotificationLogResponse(
-    @Json(name = "type")
-    val type: String,
-    @Json(name = "nickname")
-    val nickname: String,
-    @Json(name = "createdAt")
-    val createdAt: String,
-)
