--- android/app/src/main/java/com/woowacourse/ody/data/remote/entity/notification/NotificationLogsResponse.kt
@@ -1,10 +1,0 @@
-package com.woowacourse.ody.data.remote.entity.notification
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class NotificationLogsResponse(
-    @Json(name = "notiLog")
-    val logList: List<NotificationLogResponse>,
-)
