+++ android/app/src/main/java/com/woowacourse/ody/data/remote/location/repository/KakaoGeoLocationRepository.kt
@@ -1,7 +1,7 @@
 package com.woowacourse.ody.data.remote.location.repository
 
 import com.woowacourse.ody.data.remote.location.KakaoRetrofitClient
-import com.woowacourse.ody.data.remote.location.response.toGeoLocation
+import com.woowacourse.ody.data.remote.location.entity.response.toGeoLocation
 import com.woowacourse.ody.data.remote.location.service.KakaoLocationService
 import com.woowacourse.ody.domain.model.GeoLocation
 import com.woowacourse.ody.domain.repository.GeoLocationRepository
