--- android/app/src/main/java/com/woowacourse/ody/data/remote/location/response/LocationSearchResponse.kt
@@ -1,10 +1,0 @@
-package com.woowacourse.ody.data.remote.location.response
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class LocationSearchResponse(
-    @Json(name = "documents") val documents: List<Document>,
-    @Json(name = "meta") val meta: Meta,
-)
