--- android/app/src/main/java/com/woowacourse/ody/data/remote/location/response/Mapper.kt
@@ -1,7 +1,0 @@
-package com.woowacourse.ody.data.remote.location.response
-
-import com.woowacourse.ody.domain.model.GeoLocation
-
-fun LocationSearchResponse.toGeoLocation(address: String): GeoLocation {
-    return GeoLocation(address, documents[0].x, documents[0].y)
-}
