+++ android/app/src/main/java/com/woowacourse/ody/data/remote/ody/entity/meeting/response/MateResponse.kt
@@ -1,0 +1,10 @@
+package com.woowacourse.ody.data.remote.ody.entity.meeting.response
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class MateResponse(
+    @Json(name = "nickname")
+    val nickname: String,
+)
