+++ android/app/src/main/java/com/woowacourse/ody/data/remote/ody/service/MeetingService.kt
@@ -1,0 +1,28 @@
+package com.woowacourse.ody.data.remote.ody.service
+
+import com.woowacourse.ody.data.remote.ody.entity.meeting.request.MeetingRequest
+import com.woowacourse.ody.data.remote.ody.entity.meeting.response.MeetingResponse
+import com.woowacourse.ody.data.remote.ody.entity.meeting.response.MeetingsResponse
+import retrofit2.http.Body
+import retrofit2.http.GET
+import retrofit2.http.POST
+import retrofit2.http.Path
+
+interface MeetingService {
+    @GET("invite-codes/{inviteCode}/validate")
+    suspend fun getInviteCodeValidity(
+        @Path(value = "inviteCode") inviteCode: String,
+    )
+
+    @GET(MEETING_PATH)
+    suspend fun getMeeting(): MeetingsResponse
+
+    @POST("meetings")
+    suspend fun postMeeting(
+        @Body meetingRequest: MeetingRequest,
+    ): MeetingResponse
+
+    companion object {
+        const val MEETING_PATH = "/meetings/me"
+    }
+}
