+++ android/app/src/main/java/com/woowacourse/ody/data/remote/ody/service/MemberService.kt
@@ -1,0 +1,12 @@
+package com.woowacourse.ody.data.remote.ody.service
+
+import retrofit2.http.POST
+
+interface MemberService {
+    @POST(MEMBER_PATH)
+    suspend fun postMember()
+
+    companion object {
+        private const val MEMBER_PATH = "/members"
+    }
+}
